% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfcode.R
\name{elfcode_compile}
\alias{elfcode_compile}
\title{Compile an ElfCode program}
\usage{
elfcode_compile(lines, file = NULL)
}
\arguments{
\item{lines}{Lines of elfcode (one instruction per line)}

\item{file}{File to read elfcode from}
}
\description{
Compile an ElfCode program
}
