% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfcode.R
\name{elfcode_min}
\alias{elfcode_min}
\title{Convert verbose to minimal elfcode}
\usage{
elfcode_min(lines, file = NULL, output = NULL)
}
\arguments{
\item{lines}{Lines of elfcode (one instruction per line)}

\item{file}{File to read elfcode from}

\item{output}{Output destination - either a filename or
\code{NULL} to return a character vector}
}
\description{
Convert verbose to minimal elfcode
}
